Move files to src/
