--- Resources/assets/scss/atom/_gluggi-header.scss
@@ -1,47 +1,0 @@
-@import "../helpers/mixins";
-@import "../helpers/variables";
-
-.gluggi-header {
-    @include gluggi-font;
-    $font-size: 60px;
-    font-size: 0;
-    background-color: $gluggi-accent;
-    color: $gluggi-inverted-text;
-    padding: 20px 80px;
-    line-height: 1.2;
-
-    ul {
-        list-style: none;
-        display: block;
-        white-space: nowrap;
-        overflow: hidden;
-        text-overflow: ellipsis;
-        margin: 0;
-        padding: 0;
-    }
-
-    li {
-        display: inline-block;
-        font-size: $font-size;
-        opacity: .5;
-        font-weight: 200;
-
-        + li::before {
-            content: "/";
-            display: inline-block;
-            padding: 0 25px;
-            opacity: .5;
-        }
-
-        &:last-child {
-            opacity: 1;
-            cursor: default;
-        }
-    }
-
-    a {
-        text-decoration: none;
-        color: inherit;
-        cursor: pointer;
-    }
-}
