--- Resources/assets/scss/atom/_gluggi-index.scss
@@ -1,84 +1,0 @@
-@import "../helpers/mixins";
-@import "../helpers/variables";
-
-.gluggi-index {
-    $section-margin: 60px;
-    @include gluggi-font;
-    max-width: 1600px;
-    margin: 50px 80px 100px;
-
-    h1 {
-        font-size: 45px;
-        margin: 0 0 40px;
-        padding-left: 20px;
-        border-left: 4px solid $gluggi-accent;
-    }
-
-    h2 {
-        font-size: 40px;
-        margin: 0 0 20px;
-
-        a {
-            font-size: 16px;
-        }
-    }
-
-    * + h2 {
-        margin-top: $section-margin;
-    }
-
-    ul {
-        list-style: none;
-    }
-
-    li + li {
-        margin-top: 3px;
-    }
-
-    a {
-        color: $gluggi-accent;
-        text-decoration: none;
-
-        &:hover {
-            text-decoration: underline;
-        }
-    }
-
-    .gluggi-link-single {
-        margin-left: 20px;
-        font-size: 12px;
-    }
-
-    .gluggi-index-column {
-        width: 100%;
-
-        + .gluggi-index-column {
-            margin-top: $section-margin;
-        }
-    }
-
-
-    @media screen and (min-width: 900px) {
-
-        .gluggi-index-columns {
-            display: flex;
-            align-items: stretch;
-        }
-
-        .gluggi-index-column {
-            &:nth-last-child(2):first-child,
-            + .gluggi-index-column {
-                width: calc(50% - 20px);
-                flex-shrink: 0;
-                flex-grow: 0;
-            }
-
-
-            + .gluggi-index-column {
-                margin-left: 40px;
-                margin-top: 0;
-            }
-        }
-
-    }
-}
