--- Resources/assets/scss/atom/_gluggi-preview.scss
@@ -1,10 +1,0 @@
-.gluggi-preview {
-    box-sizing: border-box;
-    min-height: 80vh;
-    padding: 200px 0;
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: relative;
-}
