--- Resources/assets/scss/atom/_gluggi-varations.scss
@@ -1,5 +1,0 @@
-.gluggi-variations {
-    > * + * {
-        margin-top: 25px;
-    }
-}
