--- Resources/assets/scss/gluggi.scss
@@ -1,11 +1,0 @@
-// Base
-@import "base/globals";
-
-// Atoms
-@import "atom/gluggi-header";
-@import "atom/gluggi-index";
-@import "atom/gluggi-preview";
-@import "atom/gluggi-varations";
-
-// Templates
-@import "templates/default";
