--- Resources/assets/scss/helpers/_mixins.scss
@@ -1,3 +1,0 @@
-@mixin gluggi-font {
-    font-family: Arial, sans-serif;
-}
