--- Resources/assets/scss/helpers/_variables.scss
@@ -1,3 +1,0 @@
-$gluggi-accent: #1e47e1;
-$gluggi-default-bg: #fff;
-$gluggi-inverted-text: #fff;
