--- Resources/assets/scss/templates/_default.scss
@@ -1,5 +1,0 @@
-@import "../helpers/variables";
-
-.gluggi-preview + .gluggi-preview {
-    border-top: 3px solid $gluggi-accent;
-}
