--- Resources/public/css/gluggi.css
@@ -1,1 +1,0 @@
-body{margin:0}.gluggi-header{font-family:Arial,sans-serif;font-size:0;background-color:#1e47e1;color:#fff;padding:20px 80px;line-height:1.2}.gluggi-header ul{list-style:none;display:block;white-space:nowrap;overflow:hidden;text-overflow:ellipsis;margin:0;padding:0}.gluggi-header li{display:inline-block;font-size:60px;opacity:.5;font-weight:200}.gluggi-header li+li::before{content:"/";display:inline-block;padding:0 25px;opacity:.5}.gluggi-header li:last-child{opacity:1;cursor:default}.gluggi-header a{text-decoration:none;color:inherit;cursor:pointer}.gluggi-index{font-family:Arial,sans-serif;max-width:1600px;margin:50px 80px 100px}.gluggi-index h1{font-size:45px;margin:0 0 40px;padding-left:20px;border-left:4px solid #1e47e1}.gluggi-index h2{font-size:40px;margin:0 0 20px}.gluggi-index h2 a{font-size:16px}.gluggi-index *+h2,.gluggi-index .gluggi-index-column+.gluggi-index-column{margin-top:60px}.gluggi-index ul{list-style:none}.gluggi-index li+li{margin-top:3px}.gluggi-index a{color:#1e47e1;text-decoration:none}.gluggi-index a:hover{text-decoration:underline}.gluggi-index .gluggi-link-single{margin-left:20px;font-size:12px}.gluggi-index .gluggi-index-column{width:100%}@media screen and (min-width:900px){.gluggi-index .gluggi-index-columns{display:-ms-flexbox;display:flex;-ms-flex-align:stretch;align-items:stretch}.gluggi-index .gluggi-index-column+.gluggi-index-column,.gluggi-index .gluggi-index-column:nth-last-child(2):first-child{width:calc(50% - 20px);-ms-flex-negative:0;flex-shrink:0;-ms-flex-positive:0;flex-grow:0}.gluggi-index .gluggi-index-column+.gluggi-index-column{margin-left:40px;margin-top:0}}.gluggi-preview{box-sizing:border-box;min-height:80vh;padding:200px 0;width:100%;display:-ms-flexbox;display:flex;-ms-flex-pack:center;justify-content:center;-ms-flex-align:center;align-items:center;position:relative}.gluggi-variations>*+*{margin-top:25px}.gluggi-preview+.gluggi-preview{border-top:3px solid #1e47e1}
