+++ src/Component/ComponentConfiguration.php
