+++ src/Component/ComponentLoader.php
