+++ src/ComponentType/ComponentTypeRegistry.php
