+++ src/Exception/UnknownComponentTypeException.php
